refactor: move community session lifecycle code to dedicated module
