+++ src/deephaven_mcp/sessions/_sessions.py
@@ -38,131 +38,10 @@ from deephaven_mcp import config
 from deephaven_mcp.config._community_session import redact_community_session_config
 from deephaven_mcp.io import load_bytes
 from ._lifecycle import close_session_safely
+from ._lifecycle_community import create_session, get_session_parameters, create_session_for_worker
 from ._errors import SessionCreationError
 
 _LOGGER = logging.getLogger(__name__)
-
-
-async def create_session(**kwargs: Any) -> Session:
-    """
-    Create and return a new Deephaven Session instance in a background thread.
-    Raises SessionCreationError if session creation fails.
-    """
-    log_kwargs = redact_community_session_config(kwargs)
-    _LOGGER.info(f"Creating new Deephaven Session with config: {log_kwargs}")
-    try:
-        session = await asyncio.to_thread(Session, **kwargs)
-    except Exception as e:
-        _LOGGER.warning(
-            f"Failed to create Deephaven Session with config: {log_kwargs}: {e}"
-        )
-        raise SessionCreationError(
-            f"Failed to create Deephaven Session with config: {log_kwargs}: {e}"
-        ) from e
-    _LOGGER.info(f"Successfully created Deephaven Session: {session}")
-    return session
-
-
-async def get_session_parameters(worker_cfg: dict[str, Any]) -> dict[str, Any]:
-    """
-    Prepare and return the configuration dictionary for Deephaven Session creation.
-    Loads certificate/key files as needed (using async I/O), redacts sensitive info for logging,
-    and returns a dictionary of parameters ready to be passed to pydeephaven.Session.
-    """
-    log_cfg = redact_community_session_config(worker_cfg)
-    _LOGGER.info(f"Session configuration: {log_cfg}")
-    host = worker_cfg.get("host", None)
-    port = worker_cfg.get("port", None)
-    auth_type = worker_cfg.get("auth_type", "Anonymous")
-    auth_token = worker_cfg.get("auth_token")
-    auth_token_env_var = worker_cfg.get("auth_token_env_var")
-    if auth_token_env_var:
-        _LOGGER.info(
-            f"Attempting to read auth token from environment variable: {auth_token_env_var}"
-        )
-        token_from_env = os.getenv(auth_token_env_var)
-        if token_from_env is not None:
-            auth_token = token_from_env
-            _LOGGER.info(
-                f"Successfully read auth token from environment variable {auth_token_env_var}."
-            )
-        else:
-            auth_token = ""
-            _LOGGER.warning(
-                f"Environment variable {auth_token_env_var} specified for auth_token but not found. Using empty token."
-            )
-    elif auth_token is None:
-        auth_token = ""
-    never_timeout = worker_cfg.get("never_timeout", False)
-    session_type = worker_cfg.get("session_type", "python")
-    use_tls = worker_cfg.get("use_tls", False)
-    tls_root_certs = worker_cfg.get("tls_root_certs", None)
-    client_cert_chain = worker_cfg.get("client_cert_chain", None)
-    client_private_key = worker_cfg.get("client_private_key", None)
-    if tls_root_certs:
-        _LOGGER.info(f"Loading TLS root certs from: {worker_cfg.get('tls_root_certs')}")
-        tls_root_certs = await load_bytes(tls_root_certs)
-        _LOGGER.info("Loaded TLS root certs successfully.")
-    else:
-        _LOGGER.debug("No TLS root certs provided for session.")
-    if client_cert_chain:
-        _LOGGER.info(
-            f"Loading client cert chain from: {worker_cfg.get('client_cert_chain')}"
-        )
-        client_cert_chain = await load_bytes(client_cert_chain)
-        _LOGGER.info("Loaded client cert chain successfully.")
-    else:
-        _LOGGER.debug("No client cert chain provided for session.")
-    if client_private_key:
-        _LOGGER.info(
-            f"Loading client private key from: {worker_cfg.get('client_private_key')}"
-        )
-        client_private_key = await load_bytes(client_private_key)
-        _LOGGER.info("Loaded client private key successfully.")
-    else:
-        _LOGGER.debug("No client private key provided for session.")
-    session_config = {
-        "host": host,
-        "port": port,
-        "auth_type": auth_type,
-        "auth_token": auth_token,
-        "never_timeout": never_timeout,
-        "session_type": session_type,
-        "use_tls": use_tls,
-        "tls_root_certs": tls_root_certs,
-        "client_cert_chain": client_cert_chain,
-        "client_private_key": client_private_key,
-    }
-    log_cfg = redact_community_session_config(session_config)
-    _LOGGER.info(f"Prepared Deephaven Session config: {log_cfg}")
-    return session_config
-
-async def create_session_for_worker(config_manager: config.ConfigManager, session_name: str) -> Session:
-    """
-    Create and return a new Deephaven Session for the given worker/session name.
-
-    This helper looks up the worker configuration, gathers session parameters, logs the (redacted) config,
-    and creates the session. It does not handle caching.
-
-    Args:
-        session_name (str): The name of the worker/session.
-        config_manager: The config manager to use for config lookup.
-
-    Returns:
-        Session: A new Deephaven Session instance.
-    """
-    _LOGGER.info(f"Creating new session: {session_name}")
-    worker_cfg = await config.get_named_config(config_manager, "community_sessions", session_name)
-    session_params = await get_session_parameters(worker_cfg)
-    log_cfg = redact_community_session_config(session_params)
-    log_cfg["session_name"] = session_name
-    _LOGGER.info(
-        f"Creating new Deephaven Session with config: (worker cache key: {session_name}) {log_cfg}"
-    )
-    session = await create_session(**session_params)
-    _LOGGER.info(f"Successfully created session: {session_name}")
-    return session
-
 
 
 class SessionManager:
